🍿 move tous les composants verso dans un dossier, le fichier index est le point d'entree du composant, voir ticket #839
