+++ src/components/verso/VersoInfo.js
@@ -8,10 +8,10 @@ import { connect } from 'react-redux'
 import { withRouter } from 'react-router-dom'
 import { Icon, Logger, capitalize } from 'pass-culture-shared'
 
-import { navigationLink } from '../utils/geolocation'
-import { isRecommendationFinished } from '../helpers'
-import { selectBookables } from '../selectors/selectBookables'
-import currentRecommendationSelector from '../selectors/currentRecommendation'
+import { navigationLink } from '../../utils/geolocation'
+import { isRecommendationFinished } from '../../helpers'
+import { selectBookables } from '../../selectors/selectBookables'
+import currentRecommendationSelector from '../../selectors/currentRecommendation'
 
 class VersoInfo extends React.PureComponent {
   componentWillMount() {
